rename test_utils.py to utils.py in tests folder
