+++ tests/__init__.py
@@ -1,3 +1,5 @@
 # -*- coding: utf-8 -*-
 
 """Dynamo test package initialization."""
+
+__version__ = "0.0.0"
